Restore EventSystem object
