+++ Assets/CursorScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8f0fffed4030202429e72faa85247c7f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
